update. structure of DrugChartSlider
