+++ src/components/DrugChartSlider/components/DrugChartSliderNotification.jsx
@@ -2,8 +2,8 @@ import React from "react";
 import PropTypes from "prop-types";
 import { NotificationCarbon } from "bahmni-carbon-ui";
 import { FormattedMessage } from "react-intl";
-import { I18nProvider } from "../../features/i18n/I18nProvider";
-import "./DrugChartSliderNotification.scss";
+import { I18nProvider } from "../../../features/i18n/I18nProvider";
+import "../styles/DrugChartSliderNotification.scss";
 
 export default function DrugChartSliderNotification(props) {
   const { hostData, hostApi } = props;
