+++ src/components/DrugChartSlider/styles/DrugChartSlider.scss
