+++ src/components/DrugChartSlider/styles/DrugChartSliderNotification.scss
