--- src/utils/DrugChartSliderUtils.js
@@ -1,11 +1,0 @@
-import axios from "axios";
-import { MEDICATIONS_BASE_URL } from "../constants";
-
-export const saveMedication = async (medication) => {
-  try {
-    const response = await axios.post(MEDICATIONS_BASE_URL, medication);
-    return response;
-  } catch (error) {
-    console.error(error);
-  }
-};
